Created photonics module and implemented Fano fit
